pyproject.toml
[project]
name = "line-openai-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "utanaka2000",email = "27200337+utanaka2000@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "line-bot-sdk (>=3.16.3,<4.0.0)",
    "openai (>=1.75.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
flask = "^3.1.0"
python-dotenv = "^1.1.0"
mypy = "^1.15.0"

